+++ Wire-iOS/Sources/UserInterface/SelfProfile/AccountView.swift
@@ -300,104 +300,6 @@ extension PersonalAccountView {
     }
 }
 
-//extension TeamImageView: RoundedViewProtocol {
-//
-//    public override class var layerClass: AnyClass {
-//        return ContinuousMaskLayer.self
-//    }
-//
-//}
-
-public final class TeamImageView: UIImageView {
-
-    public enum TeamImageViewStyle {
-        case small
-        case big
-    }
-
-    public enum Content {
-        case teamImage(Data)
-        case teamName(String)
-    }
-    
-    var content: Content {
-        didSet {
-            updateImage()
-        }
-    }
-
-    private var lastLayoutBounds: CGRect = .zero
-    private let maskLayer = CALayer()
-    internal let initialLabel = UILabel()
-    public var style: TeamImageViewStyle = .small {
-        didSet {
-            switch (self.style) {
-            case .big:
-                initialLabel.font = .largeThinFont
-            case .small:
-                applySmallStyle()
-            }
-        }
-    }
-
-    func applySmallStyle() {
-        initialLabel.font = .smallSemiboldFont
-        initialLabel.textColor = .from(scheme: .textForeground, variant: .light)
-        backgroundColor = .from(scheme: .background, variant: .light)
-    }
-
-    private func updateRoundCorner() {
-        layer.cornerRadius = 4
-        clipsToBounds = true
-    }
-
-    init(content: Content) {
-        self.content = content
-        super.init(frame: .zero)
-
-        initialLabel.textAlignment = .center
-        self.addSubview(self.initialLabel)
-        self.accessibilityElements = [initialLabel]
-        
-        constrain(self, initialLabel) { selfView, initialLabel in
-            initialLabel.centerY == selfView.centerY
-            initialLabel.centerX == selfView.centerX
-        }
-        
-        maskLayer.contentsScale = UIScreen.main.scale
-        maskLayer.contentsGravity = .center
-        self.updateImage()
-
-        applySmallStyle()
-
-        updateRoundCorner()
-    }
-    
-    required public init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    override public func layoutSubviews() {
-        super.layoutSubviews()
-        
-        if !bounds.equalTo(lastLayoutBounds) {
-            lastLayoutBounds = self.bounds
-            updateRoundCorner()
-        }
-    }
-    
-    private func updateImage() {
-        switch content {
-        case .teamImage(let data):
-            image = UIImage(data: data)
-            initialLabel.text = ""
-        case .teamName(let name):
-            image = nil
-            initialLabel.text = name.first.map(String.init) ///TODO: font
-        }
-    }
-}
-
 final class TeamAccountView: BaseAccountView {
     
     public override var collapsed: Bool {
